Combine merge and extract operations with manipulate.py
